# internal/controllers/videoController.go
package controllers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/unbot2313/go-streaming-service/internal/models"
	"github.com/unbot2313/go-streaming-service/internal/services"
)

type VideoController interface {
	GetLatestVideos(c *gin.Context)
	CreateVideo(c *gin.Context)
}

// SaveVideo		godoc
// @Summary 		Save a video
// @Description 	Upload a video file along with metadata (title and description) and save it to the AWS bucket.
// @Tags 			streaming
// @Produce 		json
// @Success 		200 {object} models.VideoSwagger{}
// @Failure 		400 {object} map[string]string
// @Failure 		500 {object} map[string]string
// @Router 			/streaming/ [get]
func (vc *VideoControllerImpl) GetLatestVideos(c *gin.Context) {
	videos, err := vc.databaseVideoService.FindLatestVideos()

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, videos)
}

// SaveVideo		godoc
// @Summary 		Save a video
// @Description 	Upload a video file along with metadata (title and description) and save to the AWS bucket.
// @Tags 			streaming
// @Accept 			multipart/form-data
// @Produce 		json
// @Param 			title formData string true "Video Title"
// @Param 			description formData string false "Video Description"
// @Param 			video formData file true "Video File"
// @Success 		200 {object} models.VideoSwagger{}
// @Failure 		400 {object} map[string]string
// @Failure 		500 {object} map[string]string
// @Router 			/streaming/upload [post]
func (vc *VideoControllerImpl) CreateVideo(c *gin.Context) {

	// Recuperar el usuario del contexto
	user, exists := c.Get("user")
	if !exists {
		c.JSON(500, gin.H{"error": "User not found in context"})
		return
	}

	// Convertir a tipo User
	authenticatedUser, ok := user.(*models.User)
	if !ok {
		c.JSON(500, gin.H{"error": "Failed to parse user data"})
		return
	}

	// verificar si el archivo es válido
	if !vc.videoService.IsValidVideoExtension(c) {
		c.JSON(http.StatusBadRequest, gin.H{"error": "El archivo no es un tipo de video válido."})
		return
	}

	fileSize := c.Request.ContentLength
	const maxFileSize = 100 * 1024 * 1024 // 100 MB
	if fileSize > maxFileSize {
		c.JSON(http.StatusBadRequest, gin.H{"error": "El archivo excede el límite de tamaño permitido."})
		return
	}

	// guardar archivo en local
	videoData, err := vc.videoService.SaveVideo(c)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	// borrar el archivo original
	defer vc.videoService.GetFilesService().RemoveFile(videoData.LocalPath)

	// comprimir el video
	// pendiente

	//pasar a archivos .ts y .m3u8 con ffmpeg y guardarlo en local
	filesPath ,err := vc.videoService.FormatVideo(videoData.UniqueName)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	// borrar archivos locales .ts y .m3u8
	defer vc.videoService.GetFilesService().RemoveFolder(filesPath)



	// subir el video a s3
	savedDataInS3, err := vc.videoService.UploadFilesFromFolderToS3(filesPath)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	videoData.M3u8FileURL = savedDataInS3

	// finalmente, guardar la url del video en la base de datos
	Video, err := vc.databaseVideoService.CreateVideo(videoData, authenticatedUser.Id)
	if err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, Video)
}

type VideoControllerImpl struct {
	videoService services.VideoService;
	databaseVideoService services.DatabaseVideoService
}

func NewVideoController(videoService services.VideoService, databaseVideoService services.DatabaseVideoService) VideoController {
	return &VideoControllerImpl{
		videoService: videoService,
		databaseVideoService: databaseVideoService,
	}
}